time expanded pretty much working

+++ parser/src/main/java/TimeExpanded/NodeOrder.java
@@ -1,19 +1,14 @@
 package TimeExpanded;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.Map;
-
 /**
  * Created by bjozz on 4/8/2017.
  */
 public class NodeOrder implements Comparable{
     public String nodeId;
-    public int minute;
+    public double minute;
     public int type;
     public String stopId;
-    public NodeOrder(String nodeId, int minute, int type, String stopId) {
+    public NodeOrder(String nodeId, double minute, int type, String stopId) {
         this.nodeId = nodeId;
         this.minute = minute;
         this.type = type;
